working decision tree and put in separate component
